Process events in background with celery
