+++ bot/bot-app/events/event_handler/errors.py
@@ -1,0 +1,4 @@
+
+
+class UnknownEvent(Exception):
+    pass
